データの位置を変更（暫定的なので後で変わる可能性高い）
