+++ .idea/workspace.xml
@@ -2,12 +2,10 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="23e0e5e1-84f5-4332-bbbe-4ab183b08e4b" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/client/src/components/molecules/LockedImage.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/client/src/components/QuestionView.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/assets/stylesheets/styles.css" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/assets/stylesheets/styles.css" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/client/src/components/MainView.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/MainView.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/client/src/components/MenuListView.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/MenuListView.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/components/atoms/Image.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/atoms/Image.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/components/organisms/LockedRecipeModal.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/organisms/LockedRecipeModal.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -62,7 +60,7 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1599615597630</updated>
-      <workItem from="1599615599307" duration="13980000" />
+      <workItem from="1599615599307" duration="14724000" />
     </task>
     <servers />
   </component>
