+++ client/src/components/QuestionView.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+// My Components
+function QuestionView() {
+    return (
+        <div>TETS
+        </div>
+    );
+}
+
+export default QuestionView;
